fix: get devtools dev environment working in monorepo
