+++ .nuxtrc
@@ -1,2 +1,2 @@
 typescript.tsConfig.compilerOptions.noUncheckedIndexedAccess=true
-modules[]=../modules/devtools
+modules[]=nuxt-fonts-devtools
