+++ package.json
@@ -41,7 +41,6 @@
   },
   "dependencies": {
     "@nuxt/devtools-kit": "^1.0.8",
-    "@nuxt/devtools-ui-kit": "^1.0.8",
     "@nuxt/kit": "^3.10.3",
     "chalk": "^5.3.0",
     "css-tree": "^2.3.1",
@@ -76,6 +75,7 @@
     "execa": "^8.0.1",
     "nitropack": "^2.9.3",
     "nuxt": "^3.10.3",
+    "nuxt-fonts-devtools": "workspace:^",
     "playwright-core": "^1.42.1",
     "semver": "^7.6.0",
     "typescript": "^5.4.2",
